refactor(rubocop): remove remaining .rubocop.yml relaxations for session.rb and mapping.rb
Rename predicates (has_error_codes? -> error_codes?, has_status? -> status?), extract SessionHelpers module with pure-function helpers, refactor mqsc_command into smaller methods, and extract key-value-map handling in mapping.rb. All file-level exclusions and AllowedMethods entries removed.
