+++ sig/mq/rest/admin/sync.rbs
@@ -69,7 +69,7 @@ module MQ
         def start_and_poll: (String name, ObjectTypeConfig obj_config, SyncConfig? config) -> SyncResult
         def stop_and_poll: (String name, ObjectTypeConfig obj_config, SyncConfig? config) -> SyncResult
         def do_restart: (String name, ObjectTypeConfig obj_config, SyncConfig? config) -> SyncResult
-        def has_status?: (Array[Hash[String, untyped]] rows, Array[String] status_keys, Set[String] target_values) -> bool
+        def status?: (Array[Hash[String, untyped]] rows, Array[String] status_keys, Set[String] target_values) -> bool
         def clock_now: () -> Float
         def sleep_interval: (Float seconds) -> void
       end
